+++ basic-calculator/src/math.c
@@ -1,19 +1,19 @@
 #include "math.h"
 #include <stdio.h>
 
-int add(int a, int b) {
+float add(float a, float b) {
 	return a + b;
 }
 
-int substract(int a, int b){
+float substract(float a, float b){
 	return a - b;
 }
 
-int multiply(int a, int b){
+float multiply(float a, float b){
 	return a * b;
 }
 
-int divide(int a, int b){
+float divide(float a, float b){
 	if (b == 0) {
 		printf("Error: You can't devide by 0 value!\n");
 		return 0;
